(restruct) Move everything into a 'mybuild_embox.lang_legacy' package
