+++ mybuild_embox/__init__.py
@@ -1,0 +1,5 @@
+"""
+Embox-specific build tools.
+"""
+
+__version__ = '0.1'
